backend/requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

# LangChain + LLM
langchain==0.2.1
langchain-community==0.2.1
langchain-groq==0.1.4
langchain-huggingface==0.0.3
groq==0.9.0

# Vector Store + Embeddings
faiss-cpu==1.13.2
sentence-transformers==3.0.0
rank-bm25==0.2.2

# Hugging Face
transformers==4.41.2
torch==2.9.0
huggingface-hub==0.23.2
datasets==2.19.1

# Database
motor==3.4.0
pymongo==4.7.2
redis==5.0.4
beanie==1.26.0

# WhatsApp
twilio==9.1.0

# Document Processing
pypdf==4.2.0
python-docx==1.1.2
reportlab==4.2.0
Pillow==10.4.0

# Translation + NLP
deep-translator==1.11.4
langdetect==1.0.9
indic-transliteration==2.3.56

# Auth + Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9

# HTTP + Scraping
httpx==0.27.0
aiohttp==3.9.5
beautifulsoup4==4.12.3
playwright==1.44.0
lxml==5.2.2

# Background Tasks
celery==5.4.0
apscheduler==3.10.4

# Monitoring
loguru==0.7.2
sentry-sdk[fastapi]==2.3.1

# Utils
tenacity==8.3.0
cachetools==5.3.3
python-slugify==8.0.4
